A sector holds ten thousand, and a quantity stops being a short
Issue #91. Richard 2026-09-10: "since everything was moved to a short, there was no need
to shorten that to 1000. Go ahead and make that 10,000."

The original's 1000 was a consequence of packing quantities into ten bits, not a design
choice. We are not packing, so the cap is ours to pick. `default_capacity` is 10,000.

The catch is that a warehouse and a city carry store_multiplier 10, so the real ceiling
is 100,000 and a short stops at 32,767. Left as shorts a warehouse would clamp at a
third of its stated capacity — and not quietly for long: the apply step computes the new
stock as an int, so the clamp would surface as the conservation check finding the books
short and throwing. Quantities are ints now: a sector's stock, a ship's hold, a
threshold and a delivery threshold. Ledger.stock was already int and the tallies already
long, so nothing there moved.

It costs 72 bytes on a sector carrying stock, thresholds and delivery orders at once —
272 to 344, against 610 before #77, and only on sectors that hold something, since an
empty one still shares the canonical zero. That is a good trade for deleting a ceiling
that was sitting in the rules pretending to be a design decision.

CapacityTest is the test that would have caught leaving them short. Nothing in the suite
had ever reached 32,767 of anything, so nothing would have noticed: it puts 100,000 iron
in a warehouse, runs an update, and checks it is all still there — and that 150,000
still spoils down to the cap, because the cap is still a cap.

The golden hash moved, as it must: capacity is a rule and this changes it.

+++ empire-engine/src/main/java/org/hastingtx/empire/engine/model/DeliverOrders.java
@@ -8,18 +8,18 @@ import java.util.concurrent.ConcurrentHashMap;
  * Above {@code threshold[c]}, whatever is there moves one hex in direction {@code dir[c]} at the
  * update; {@code NONE} means no order. Directions index {@link org.hastingtx.empire.engine.geo.Hex#DIR_NAMES}.
  *
- * <p>Held narrow (issue #77): a direction is one of six, so a byte; a threshold is a quantity, so a
- * short, exact — byte scaling would make {@code deliver food 100} mean something else. 228 bytes a
- * sector becomes about 100, and a sector with no orders at all — most of a large world — shares the
+ * <p>Held narrow (issue #77): a direction is one of six, so a byte; a threshold is a quantity, so an
+ * int, exact — byte scaling would make {@code deliver food 100} mean something else, and a short would
+ * stop short of a warehouse's 100,000 (issue #91). 228 bytes a sector becomes about 150, and a sector with no orders at all — most of a large world — shares the
  * canonical empty instance and costs only the reference.
  */
 public final class DeliverOrders {
     public static final int NONE = -1;
 
     private final byte[] dir;
-    private final short[] threshold;
+    private final int[] threshold;
 
-    private DeliverOrders(byte[] dir, short[] threshold) { this.dir = dir; this.threshold = threshold; }
+    private DeliverOrders(byte[] dir, int[] threshold) { this.dir = dir; this.threshold = threshold; }
 
     private static final ConcurrentHashMap<Integer, DeliverOrders> EMPTY = new ConcurrentHashMap<>();
 
@@ -28,7 +28,7 @@ public final class DeliverOrders {
         return EMPTY.computeIfAbsent(n, k -> {
             byte[] d = new byte[k];
             Arrays.fill(d, (byte) NONE);
-            return new DeliverOrders(d, new short[k]);
+            return new DeliverOrders(d, new int[k]);
         });
     }
 
@@ -42,7 +42,7 @@ public final class DeliverOrders {
 
     public DeliverOrders with(int c, int d, double thr) {
         byte[] nd = dir.clone();
-        short[] nt = threshold.clone();
+        int[] nt = threshold.clone();
         nd[c] = (byte) d;
         nt[c] = d == NONE ? 0 : Sector.thresholdOf(thr);
         for (byte v : nd) if (v != NONE) return new DeliverOrders(nd, nt);
